requirements.txt
flask==2.3.3
tldextract==5.1.1
requests==2.31.0
beautifulsoup4==4.12.3
whois==0.9.27
validators==0.22.0
python-Levenshtein==0.25.0
python-whois>=0.8.0
urllib3>=1.26.7
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
